(UNRELATED) Separating out some FSUI components...
In future we can create new dir for FilteringTab and move
related things into there.
